warningの訂正

+++ input.cpp
@@ -554,7 +554,7 @@ float CInput::GetTimeUpToReactionKey(int nNum)
 
 			if (m_fTimeEnd[0] == 0.0f)
 			{
-				m_fTimeEnd[0] = timeGetTime();
+				m_fTimeEnd[0] = (float)timeGetTime();
 			}
 			return  m_fTimeEnd[0] - m_fTimeBeginning;
 		}
@@ -573,7 +573,7 @@ float CInput::GetTimeUpToReactionKey(int nNum)
 
 			if (m_fTimeEnd[1] == 0.0f)
 			{
-				m_fTimeEnd[1] = timeGetTime();
+				m_fTimeEnd[1] = (float)timeGetTime();
 			}
 			return  m_fTimeEnd[1] - m_fTimeBeginning;
 		}
@@ -588,7 +588,7 @@ float CInput::GetTimeUpToReactionKey(int nNum)
 void CInput::TimeUpToReactionKeyTimeStart()
 {
 	m_bTimeStart = true;
-	m_fTimeBeginning = timeGetTime();
+	m_fTimeBeginning = (float)timeGetTime();
 	m_fTimeEnd[0] = 0.0f;
 	m_fTimeEnd[1] = 0.0f;
